+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## v0.0.7 - 2/28/23
-- Added PlatformApiClient for oauth2 authenticated calls to our Platform API
+- Added Oauth2ApiClient for oauth2 authenticated calls to our Platform API and Sierra
 
 ## v0.0.5 - 2/22/23
 - Support write queries to PostgreSQL and MySQL databases